Feature: Save and get a link (#3)
Document the coverage command

Use an interface for the LinkAdapter: The interface is explicit, unlike duck typing. Developers know what the implementation should do.

The domain should have the knowledge of the LinkAdapter interface, but not any knowledge of the implementations of this interface.

Keep it simple: a DAO pattern is too generic. We don't need that for the moment, we only want to store Link entities with a LinkAdapter.

Structure the files according to the hexagonal architecture: LinkAdapter interface in the domain, implementations in the infrastructure.

The implementations of this interface have the knowledge of the LinkAdapter interface they should implement and the entities of the domain.

Interfaces are excluded from the coverage report, because they don't implement any logic.
The faker module is now used during the tests.